CS172 – Socialites
This was the first homework assignment that I created for my CS 172 class.  In this program, a user can create a pseudo-social network page called a Socalite.  The user has to enter the information that is asked by the console and then a text file with the supplied information will be created along with an HTML file that contains the Socialite itself.  Five pre-made Socalites are provided in this folder.